chore: rename DAG related modules
